server/ai_models/chatbot/requirements.txt
google-generativeai>=0.8.2
python-dotenv>=1.0.1
